refactor(triage): unify pr_is_merged onto single impl, extract gateway (#4624)
CliGitHubGateway.pr_is_merged and commit_exists now delegate to
verify_pr_merged and verify_commit_exists from verify_issue_close.py,
removing the divergent copies that lacked the ancestry check (issue 4624
batch path). Extracted CliGitHubGateway and OfflineGateway into
scripts/triage_gateway.py to bring triage_batch_apply.py under 500 lines
(491 from 659). Gateway tests and scope wiring tests extracted into
tests/test_triage_gateway.py (295 lines). Added pinning tests for
open-PR rejection and ancestry failure through the batch path.
